backend/pyproject.toml
[project]
name = "agent-hunt-backend"
version = "0.1.0"
description = "AI Agent job market analysis — backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.30,<1",
    "sqlalchemy[asyncio]>=2.0,<3",
    "asyncpg>=0.30,<1",
    "alembic>=1.14,<2",
    "pydantic>=2.0,<3",
    "pydantic-settings>=2.0,<3",
    "redis>=5.0,<6",
    "celery[redis]>=5.4,<6",
    "httpx>=0.27,<1",
    "google-genai>=1.0,<2",
    "pgvector>=0.3,<1",
    "python-dotenv>=1.0,<2",
    "python-jobspy>=1.1,<2",
    "playwright>=1.40,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9",
    "pytest-asyncio>=0.24,<1",
    "httpx>=0.27,<1",
    "ruff>=0.8,<1",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
